Ya no aparece el error 'Error al eliminar el usuario' al volver al listado de usuarios luego de crear o actualizar un usuario

+++ src/pages/Admin/AdminUsers/AdminUsersList/AdminUsersList.jsx
@@ -49,9 +49,10 @@ const AdminUsersList = () => {
   },[status?.code])
   
   useEffect(() => {
-    if(modal.isOkClicked){  //Se seleccionó el botón de aceptar en el modal
+    if(modal.isOkClicked && deleteId){  //Se seleccionó el botón de aceptar en el modal
       dispatch(deleteUser(deleteId, access))
       setEliminado(true)
+      setDeleteId(null)
     }
   },[modal])
